Updated Spanish translation
I'm not sure if there's an issue about this update/request. Hope that I'm not causing any troubles, thank you in advance guys.

As you must know, the es.yml file contains most of the content for the getmonero dot org site. There are other files that needs translation and I'll be working on them too. I experienced just one problem when testing the file: the Stack Exchange section didn't translated, which corresponds to line 104. Maybe that line depends on other file in the es folder but I'm not aware of that now. 

A minor problem is the space used when translating some of the titles in the Team page (Research Lab, which translates into Laboratorio de Investigación, actually I just put "Lab. de Investigación" but it's still too long, and Special Thanks - Agradecimientos Especiales. Those titles look like they need more space to fit it.